Added explanatory comment
The app was supposed to integrate with Venmo, but when it was being set
up, Venmo refused to authorize any of our teammates so none of us could
get the necessary API key to use in our app. Because of this, we
intended to switch to another service but ran out of time.